requirements.txt
beautifulsoup4==4.12.3
boto3==1.35.97
botocore==1.35.97
certifi==2024.12.14
charset-normalizer==3.4.1
confluent-kafka==2.8.0
contourpy==1.3.1
cycler==0.12.1
fonttools==4.55.3
frozendict==2.4.6
geopandas==1.0.1
html5lib==1.1
idna==3.10
jmespath==1.0.1
kafka-python==2.0.2
kiwisolver==1.4.8
lxml==5.3.0
matplotlib==3.10.0
multitasking==0.0.11
networkx==3.4.2
numpy==2.2.1
osmnx==2.0.1
packaging==24.2
pandas==2.2.3
peewee==3.17.8
pillow==11.1.0
platformdirs==4.3.6
pyogrio==0.10.0
pyparsing==3.2.1
pyproj==3.7.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
redis==5.2.1
requests==2.32.3
s3transfer==0.10.4
shapely==2.0.6
six==1.17.0
soupsieve==2.6
tzdata==2024.2
urllib3==2.3.0
webencodings==0.5.1
yfinance==0.2.51
